Update README.md
I created the Starbucks Corporation website to inform internet users of the different business activities it engages in. The letter is being sent by Starbucks' marketing division, which utilizes the page to explain to customers the company's goals for the market. Any consumer interested in learning more about the company through the website is the message's presumed recipient. In order to help consumers and other stakeholders understand the company's place in the industry, the "about us" section gives the company the opportunity to engage with them. The page includes details on product categories, business operations, investor relations, and news stories showcasing various events the company has taken part in. The web page features a variety of media, which enhances its appeal to users who browse its many sections. To help the intended audience readily understand the idea being given, every sentence written on the website's various pages for being concise, accurate, and clear. Pictures are present in some places on the page, which gives the information being presented more color. In order to make a good first impression on visitors, photographs are also employed to illustrate various written statements. Visitors are enticed to read more on the website by additional visual features, and this helps them form favorable opinions of the company's place in the sector. Additionally, multiple font styles that are exhibited in both muted and bold hues entice web visitors to investigate other elements on the page.